+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 _release-date_
 
 * **Breaking:** Merge `BaseLithoStartupLogging` abstract class, `LithoStartupLoggerUtil` helper class and `LithoStartupLogger` interface into single `LithoStartupLogger` abstract class.
+* **Breaking:** Consolidate two layout calculation `PerfEvent`s into one: remove `FrameworkLogEvents.EVENT_LAYOUT_CALCULATE` and move some of its annotations to `FrameworkLogEvents.CALCULATE_LAYOUT_STATE` which will be used instead.
 
 
 For more details, see the [full diff](https://github.com/facebook/litho/compare/v0.33.0...master).